move controller to claim
